third_party/blink/web_tests/external/wpt/webrtc-stats/supported-stats.https-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] inbound-rtp's framesRendered
  assert_true: Is framesRendered present expected true got false
[FAIL] inbound-rtp's fecBytesReceived
  assert_true: Is fecBytesReceived present expected true got false
[FAIL] inbound-rtp's estimatedPlayoutTimestamp
  assert_true: Is estimatedPlayoutTimestamp present expected true got false
[PRECONDITION_FAILED] inbound-rtp's fecSsrc
  inbound-rtp.fecSsrc marked as not testable.
[PRECONDITION_FAILED] outbound-rtp's rid
  outbound-rtp.rid marked as not testable.
[FAIL] remote-inbound-rtp's packetsReceived
  assert_true: Is packetsReceived present expected true got false
[FAIL] remote-outbound-rtp's roundTripTime
  assert_true: Is roundTripTime present expected true got false
[PRECONDITION_FAILED] media-source's echoReturnLoss
  media-source.echoReturnLoss marked as not testable.
[PRECONDITION_FAILED] media-source's echoReturnLossEnhancement
  media-source.echoReturnLossEnhancement marked as not testable.
[PRECONDITION_FAILED] candidate-pair's availableIncomingBitrate
  candidate-pair.availableIncomingBitrate marked as not testable.
[PRECONDITION_FAILED] local-candidate's url
  local-candidate.url marked as not testable.
[PRECONDITION_FAILED] local-candidate's relayProtocol
  local-candidate.relayProtocol marked as not testable.
[PRECONDITION_FAILED] local-candidate's relatedAddress
  local-candidate.relatedAddress marked as not testable.
[PRECONDITION_FAILED] local-candidate's relatedPort
  local-candidate.relatedPort marked as not testable.
[PRECONDITION_FAILED] remote-candidate's url
  remote-candidate.url marked as not testable.
[PRECONDITION_FAILED] remote-candidate's relayProtocol
  remote-candidate.relayProtocol marked as not testable.
[PRECONDITION_FAILED] remote-candidate's relatedAddress
  remote-candidate.relatedAddress marked as not testable.
[PRECONDITION_FAILED] remote-candidate's relatedPort
  remote-candidate.relatedPort marked as not testable.
[PRECONDITION_FAILED] remote-candidate's tcpType
  remote-candidate.tcpType marked as not testable.
[PRECONDITION_FAILED] certificate's issuerCertificateId
  certificate.issuerCertificateId marked as not testable.
Harness: the test ran to completion.
